+++ marvin/input.go
@@ -9,7 +9,8 @@ import (
 
 // https://golangdocs.com/command-line-arguments-in-golang
 type CmdArgs struct {
-	DangerousRegen string // true | false
+	DangerousRegen string // true | false	( default = false )
+	FetchOnly      string // true | false	( default = false )
 	ConfigPath     string // if no params are  specified, get the config from current dir
 }